requirements.txt
appnope==0.1.0
decorator==4.0.11
Django==1.10.6
django-cors-headers==2.0.2
django-extensions==1.7.7
django-rest-auth==0.9.1
djangorestframework==3.6.2
ipython==5.3.0
ipython-genutils==0.1.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.7
ptyprocess==0.5.1
Pygments==2.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
